Cleaned up, fixed bugs
Reorganised some of the files on the client-side and massively cleaned up the code on both sides. The join algorithm has been improved and some small bugs in it have been patched. Framework has been set such that it should be ready for more types of packets and the rest of the game.